refactor(cli): split monolithic cli.py into structured package
cli.py (763 lines) → cli/ package:
  _shared.py        app, helpers, constants
  _rendering.py     progress bar, color utilities
  indexing/         init, walk, sync, enrich commands
  insights/         stats, search, profile, analysis commands
  maintenance.py    reset, restore, install-skill commands

SQL stays co-located in the module that uses it.
Test patches updated to target sourcemap_indexer.cli.indexing.enrich.run_enrich.
pre-commit.sh: exclude deleted files from linter input (--diff-filter=d).
