+++ src/sourcemap_indexer/cli/__init__.py
@@ -1,0 +1,20 @@
+from __future__ import annotations
+
+from sourcemap_indexer.cli._rendering import _lang_color, _proportional_width
+from sourcemap_indexer.cli._shared import app
+from sourcemap_indexer.cli.indexing import enrich as _enrich_mod  # noqa: F401
+from sourcemap_indexer.cli.indexing import init as _init_mod  # noqa: F401
+from sourcemap_indexer.cli.indexing import sync as _sync_mod  # noqa: F401
+from sourcemap_indexer.cli.indexing import walk as _walk_mod  # noqa: F401
+from sourcemap_indexer.cli.insights import analysis as _analysis_mod  # noqa: F401
+from sourcemap_indexer.cli.insights import profile as _profile_mod  # noqa: F401
+from sourcemap_indexer.cli.insights import search as _search_mod  # noqa: F401
+from sourcemap_indexer.cli.insights import stats as _stats_mod  # noqa: F401
+from sourcemap_indexer.cli.maintenance import (  # noqa: F401
+    install_skill,
+    reset,
+    restore,
+)
+from sourcemap_indexer.infra.llama_client import LlamaClient
+
+__all__ = ["LlamaClient", "_lang_color", "_proportional_width", "app"]
